Vista Default Ingredientes

+++ Datos/Entidades/IngredienteEntidad.cs
@@ -10,7 +10,7 @@ namespace Datos.Entidades
     {
         public int id_ingrediente { get; set; }
         public string nombre { get; set; }
-        public float cantidad { get; set; }
+        public double cantidad { get; set; }
         public decimal costo { get; set; }
         public string proveedor { get; set; }
         public UnidadMedidaEntidad unidad { get; set; }
